% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ESPEN_api}
\alias{print.ESPEN_api}
\title{Print an ESPEN API response object}
\usage{
\method{print}{ESPEN_api}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ESPEN_api}.}

\item{...}{Ignored.}
}
\description{
Print an ESPEN API response object
}
