% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPEN_API_data.R
\name{ESPEN_API_data}
\alias{ESPEN_API_data}
\title{A function to request data from the ESPEN portal using API}
\usage{
ESPEN_API_data(
  api_key = NULL,
  country = NULL,
  iso2 = NULL,
  disease = "sth",
  level = "sitelevel",
  type = FALSE,
  subtype = "mda",
  start_year = NULL,
  end_year = NULL,
  limit = NULL,
  offset = NULL,
  attributes = NULL,
  df = TRUE
)
}
\arguments{
\item{api_key}{the API key from ESPEN. You can request the key from ESPEN portal. To use the default API key, ignore the argument or set api_key = NULL.}

\item{country}{the name of the country. E.g `Ghana`}

\item{iso2}{optional: country ISO2 code, e.g. `GH`. you can either specify a country name or the ISO2 code}

\item{disease}{specify the disease of interest, options are lf , oncho , loa , sch , sth , trachoma , coendemicity}

\item{level}{specify which the level of the data, options are `iu` and  `sitelevel`. iu means implemantation unit.}

\item{type}{logical: should you want a forcast of MDA or impact assessment. default is FALSE.}

\item{subtype}{only use when type =TRUE, options are mda and impact_assessment.}

\item{start_year}{specify the starting year of the data, e.g start_year = 2010}

\item{end_year}{specify the starting year of the data, e.g end_year = 2020}

\item{limit}{to specify how many records to return}

\item{offset}{to specify how many records to skip}

\item{attributes}{options are IU_ID, Endemicity, MDA, EffMDA. This is basically used to reduce the data.}

\item{df}{logical: should the function return a dataframe (default) or JSON.}
}
\value{
A dataframe or JSON file of the data
}
\description{
Allows users to request data for countries through the ESPEN API.  Users provide parameters as specified in <\url{https://admin.espen.afro.who.int/docs/api/data}> or as described below and the function returns a JSON string or data frame.
}
\details{
See <\url{https://espen.afro.who.int/tools-resources/download-data}>  for more details on the data available.
}
\examples{
library(ESPENAPI)
# extract the STH data from Nigeria at site level from year 2010 to 2020.
data  <- ESPEN_API_data(country="Nigeria", disease="sth",
level="sitelevel", start_year=2010, end_year=2010)
head(data)
}
